requirements.txt
aiohttp==3.7.4.post0
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
braceexpand==0.1.6
chardet==4.0.0
discord.py==1.7.2
humanize==3.5.0
idna==3.1
import-expression==1.1.4
jishaku==2.0.0
lxml==4.6.3
multidict==5.1.0
psutil==5.8.0
six==1.15.0
soupsieve==2.2.1
toml==0.10.2
typing-extensions==3.10.0.0
yarl==1.6.3
